Refactor : 태그 Id 검색으로 수정

+++ server/src/main/java/com/singsong/db/entity/Tag.java
@@ -17,7 +17,8 @@ import javax.persistence.*;
 public class Tag {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    Long tagId;
+    int tagId;
+
     @Column(length = 20)
     String tagName;
 }
