+++ server/src/main/java/com/singsong/db/repository/TagRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface TagRepository  extends JpaRepository<Tag, Long> {
+public interface TagRepository  extends JpaRepository<Tag, Integer> {
     Optional<Tag> findByTagName(String tagName);
 }
